lab_5/web_app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
google-genai>=0.3.0
googlemaps>=4.10.0
requests>=2.31.0
google-cloud-modelarmor>=0.1.0
google-api-core>=2.15.0
